requirements.txt
fastapi
uvicorn
requests
python-dotenv

faiss-cpu
sentence-transformers

langchain-community
langchain-text-splitters
langchain-huggingface
